% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadershipdata.R
\docType{data}
\name{leader_cult}
\alias{leader_cult}
\title{Culture-level data from HRAF variables}
\format{A data frame with 59 rows and 68 variables:
\describe{
  \item{\code{c_name}}{character. DESCRIPTION.}
  \item{\code{neel_better.mates_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{neel_big.family_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{neel_intelligence_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{neel_polygynous_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_counsel_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_emulated_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_expertise_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_family_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_f.exp.success_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_likable_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{prestige_respected_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_aggression_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_assert.authority_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_avoid.dom_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_fear_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_fighting_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_personality_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_reputation_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{dom_strong_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{hooper_performance_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{hooper_sanction.freeriders_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{hooper_payoff_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{hooper_group.size_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{hooper_coop.activities_present}}{double. Value of 1 indicates as least one codesheeet/text provides evidence for this variable, for this culture.}
  \item{\code{neel_better.mates_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{neel_big.family_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{neel_intelligence_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{neel_polygynous_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_counsel_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_emulated_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_expertise_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_family_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_f.exp.success_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_likable_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{prestige_respected_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_aggression_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_assert.authority_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_avoid.dom_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_fear_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_fighting_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_personality_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_reputation_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{dom_strong_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{hooper_performance_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{hooper_sanction.freeriders_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{hooper_coop.activities_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{hooper_payoff_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{hooper_group.size_mean}}{double. Mean value of variable across all codesheets for each culture. Values greater than 0 indicate greater evidence for than evidence against.}
  \item{\code{c_culture_code}}{integer. HRAF OWC culture code.}
  \item{\code{Location}}{integer. Nation state which culture is located within, from the SCCS}
  \item{\code{Documents}}{integer. Number of eHRAF documents providing data for that culture.}
  \item{\code{V158.1}}{integer. SCCS cultural complexity variable}
  \item{\code{c_cultural_complexity}}{integer. SCCS cultural complexity value (SCCS V158.1), with modifications *check discrepancies}
  \item{\code{c_latitude}}{integer. Location latitude.}
  \item{\code{c_longitude}}{integer. Location longitude.}
  \item{\code{c_region}}{character. HRAF region}
  \item{\code{c_subregion}}{character. HRAF subregion}
  \item{\code{c_subsistence_code}}{integer. Numeric code assigned to HRAF subsistence classification.}
  \item{\code{c_subsistence_type}}{character. HRAF subsistence classification.}
  \item{\code{extract_count}}{integer. Number of texts associated with culture.}
  \item{\code{neel_total_cell_count}}{integer. Total number of cells associated with the neel model in the text x variable matrix for each culture. }
  \item{\code{neel_total_for}}{double. DESCRIPTION.}
  \item{\code{neel_total_none}}{double. DESCRIPTION.}
  \item{\code{neel_total_against}}{double. DESCRIPTION.}
  \item{\code{prest_total_cell_count}}{integer. Total number of cells associated with the prestige model in the text x variable matrix for each culture.}
  \item{\code{prest_total_for}}{double. DESCRIPTION.}
  \item{\code{prest_total_none}}{double. DESCRIPTION.}
  \item{\code{prest_total_against}}{double. DESCRIPTION.}
  \item{\code{dom_total_cell_count}}{integer. Total number of cells associated with the dominance model in the text x variable matrix for each culture.}
  \item{\code{dom_total_for}}{double. DESCRIPTION.}
  \item{\code{dom_total_none}}{double. DESCRIPTION.}
  \item{\code{dom_total_against}}{double. DESCRIPTION.}
  \item{\code{hooper_total_for}}{double. DESCRIPTION.}
  \item{\code{hooper_total_none}}{double. DESCRIPTION.}
  \item{\code{hooper_total_against}}{double. DESCRIPTION.}
  \item{\code{subsistence}}{integer. DESCRIPTION.}
  \item{\code{region}}{character. DESCRIPTION.}
  \item{\code{settlement_fixity}}{integer. DESCRIPTION.}
  \item{\code{com_size}}{integer. DESCRIPTION.}
  \item{\code{pop_density}}{integer. DESCRIPTION.}
  \item{\code{neel_cult_score}}{double. DESCRIPTION.}
  \item{\code{prest_cult_score}}{double. DESCRIPTION.}
  \item{\code{dom_cult_score}}{double. DESCRIPTION.}
}}
\usage{
leader_cult
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
