% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadershipdata.R
\docType{data}
\name{leader_text}
\alias{leader_text}
\title{Text-level data from HRAF variables}
\format{A data frame with 1212 rows and 51 variables:
\describe{
  \item{\code{cs_ID}}{integer. DESCRIPTION.}
  \item{\code{cs_textrec_ID}}{double. DESCRIPTION.}
  \item{\code{text_type}}{integer. DESCRIPTION.}
  \item{\code{c_culture_code}}{integer. DESCRIPTION.}
  \item{\code{c_name}}{integer. DESCRIPTION.}
  \item{\code{c_region}}{integer. DESCRIPTION.}
  \item{\code{c_subregion}}{integer. DESCRIPTION.}
  \item{\code{c_subsistence_type}}{integer. DESCRIPTION.}
  \item{\code{c_subsistence_code}}{integer. DESCRIPTION.}
  \item{\code{demo_sex}}{integer. DESCRIPTION.}
  \item{\code{dom_aggression}}{double. DESCRIPTION.}
  \item{\code{dom_assert.authority}}{double. DESCRIPTION.}
  \item{\code{dom_avoid.dom}}{double. DESCRIPTION.}
  \item{\code{dom_fear}}{double. DESCRIPTION.}
  \item{\code{dom_fighting}}{double. DESCRIPTION.}
  \item{\code{dom_personality}}{double. DESCRIPTION.}
  \item{\code{dom_reputation}}{double. DESCRIPTION.}
  \item{\code{dom_strong}}{double. DESCRIPTION.}
  \item{\code{neel_better.mates}}{double. DESCRIPTION.}
  \item{\code{neel_big.family}}{double. DESCRIPTION.}
  \item{\code{neel_intelligence}}{double. DESCRIPTION.}
  \item{\code{neel_polygynous}}{double. DESCRIPTION.}
  \item{\code{prestige_counsel}}{double. DESCRIPTION.}
  \item{\code{prestige_emulated}}{double. DESCRIPTION.}
  \item{\code{prestige_expertise}}{double. DESCRIPTION.}
  \item{\code{prestige_family}}{double. DESCRIPTION.}
  \item{\code{prestige_f.exp.success}}{double. DESCRIPTION.}
  \item{\code{prestige_likable}}{double. DESCRIPTION.}
  \item{\code{prestige_respected}}{double. DESCRIPTION.}
  \item{\code{evidence_prestige_for}}{double. DESCRIPTION.}
  \item{\code{evidence_prestige_against}}{double. DESCRIPTION.}
  \item{\code{evidence_dom_for}}{double. DESCRIPTION.}
  \item{\code{evidence_dom_against}}{double. DESCRIPTION.}
  \item{\code{evidence_neel_for}}{double. DESCRIPTION.}
  \item{\code{evidence_neel_against}}{double. DESCRIPTION.}
  \item{\code{subsistence2}}{integer. DESCRIPTION.}
  \item{\code{region2}}{integer. DESCRIPTION.}
  \item{\code{dom_for}}{double. DESCRIPTION.}
  \item{\code{dom_against}}{double. DESCRIPTION.}
  \item{\code{dom_none}}{double. DESCRIPTION.}
  \item{\code{dom_sum}}{double. DESCRIPTION.}
  \item{\code{prestige_for}}{double. DESCRIPTION.}
  \item{\code{prestige_against}}{double. DESCRIPTION.}
  \item{\code{prestige_none}}{double. DESCRIPTION.}
  \item{\code{prestige_sum}}{double. DESCRIPTION.}
  \item{\code{neel_for}}{double. DESCRIPTION.}
  \item{\code{neel_against}}{double. DESCRIPTION.}
  \item{\code{neel_none}}{double. DESCRIPTION.}
  \item{\code{neel_sum}}{double. DESCRIPTION.}
  \item{\code{dom_totals}}{double. DESCRIPTION.}
  \item{\code{prest_totals}}{double. DESCRIPTION.}
}}
\source{
\url{http://www.ehrafsite.info/}
}
\usage{
leader_text
}
\description{
1200+ ethnographic text coded as evidence for, against, no evidence on variables from theories of leadership.
}
\keyword{datasets}

