% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadershipdata.R
\docType{data}
\name{leader_cult}
\alias{leader_cult}
\title{Culture-level data from HRAF variables}
\format{A data frame with 59 rows and 68 variables:
\describe{
  \item{\code{c_name}}{character. DESCRIPTION.}
  \item{\code{neel_better.mates_present}}{double. DESCRIPTION.}
  \item{\code{neel_big.family_present}}{double. DESCRIPTION.}
  \item{\code{neel_intelligence_present}}{double. DESCRIPTION.}
  \item{\code{neel_polygynous_present}}{double. DESCRIPTION.}
  \item{\code{prestige_counsel_present}}{double. DESCRIPTION.}
  \item{\code{prestige_emulated_present}}{double. DESCRIPTION.}
  \item{\code{prestige_expertise_present}}{double. DESCRIPTION.}
  \item{\code{prestige_family_present}}{double. DESCRIPTION.}
  \item{\code{prestige_f.exp.success_present}}{double. DESCRIPTION.}
  \item{\code{prestige_likable_present}}{double. DESCRIPTION.}
  \item{\code{prestige_respected_present}}{double. DESCRIPTION.}
  \item{\code{dom_aggression_present}}{double. DESCRIPTION.}
  \item{\code{dom_assert.authority_present}}{double. DESCRIPTION.}
  \item{\code{dom_avoid.dom_present}}{double. DESCRIPTION.}
  \item{\code{dom_fear_present}}{double. DESCRIPTION.}
  \item{\code{dom_fighting_present}}{double. DESCRIPTION.}
  \item{\code{dom_personality_present}}{double. DESCRIPTION.}
  \item{\code{dom_reputation_present}}{double. DESCRIPTION.}
  \item{\code{dom_strong_present}}{double. DESCRIPTION.}
  \item{\code{neel_better.mates_mean}}{double. DESCRIPTION.}
  \item{\code{neel_big.family_mean}}{double. DESCRIPTION.}
  \item{\code{neel_intelligence_mean}}{double. DESCRIPTION.}
  \item{\code{neel_polygynous_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_counsel_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_emulated_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_expertise_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_family_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_f.exp.success_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_likable_mean}}{double. DESCRIPTION.}
  \item{\code{prestige_respected_mean}}{double. DESCRIPTION.}
  \item{\code{dom_aggression_mean}}{double. DESCRIPTION.}
  \item{\code{dom_assert.authority_mean}}{double. DESCRIPTION.}
  \item{\code{dom_avoid.dom_mean}}{double. DESCRIPTION.}
  \item{\code{dom_fear_mean}}{double. DESCRIPTION.}
  \item{\code{dom_fighting_mean}}{double. DESCRIPTION.}
  \item{\code{dom_personality_mean}}{double. DESCRIPTION.}
  \item{\code{dom_reputation_mean}}{double. DESCRIPTION.}
  \item{\code{dom_strong_mean}}{double. DESCRIPTION.}
  \item{\code{c_culture_code}}{integer. DESCRIPTION.}
  \item{\code{Location}}{integer. DESCRIPTION.}
  \item{\code{Position}}{integer. DESCRIPTION.}
  \item{\code{c_cultural_complexity}}{integer. DESCRIPTION.}
  \item{\code{c_region}}{character. DESCRIPTION.}
  \item{\code{c_subregion}}{character. DESCRIPTION.}
  \item{\code{c_subsistence_code}}{integer. DESCRIPTION.}
  \item{\code{c_subsistence_type}}{character. DESCRIPTION.}
  \item{\code{extract_count}}{integer. DESCRIPTION.}
  \item{\code{neel_total_cell_count}}{integer. DESCRIPTION.}
  \item{\code{neel_total_for}}{double. DESCRIPTION.}
  \item{\code{neel_total_none}}{double. DESCRIPTION.}
  \item{\code{neel_total_against}}{double. DESCRIPTION.}
  \item{\code{prest_total_cell_count}}{integer. DESCRIPTION.}
  \item{\code{prest_total_for}}{double. DESCRIPTION.}
  \item{\code{prest_total_none}}{double. DESCRIPTION.}
  \item{\code{prest_total_against}}{double. DESCRIPTION.}
  \item{\code{dom_total_cell_count}}{integer. DESCRIPTION.}
  \item{\code{dom_total_for}}{double. DESCRIPTION.}
  \item{\code{dom_total_none}}{double. DESCRIPTION.}
  \item{\code{dom_total_against}}{double. DESCRIPTION.}
  \item{\code{subsistence2}}{integer. DESCRIPTION.}
  \item{\code{region2}}{character. DESCRIPTION.}
  \item{\code{settlement_fixity2}}{integer. DESCRIPTION.}
  \item{\code{com_size2}}{integer. DESCRIPTION.}
  \item{\code{pop_density2}}{integer. DESCRIPTION.}
  \item{\code{neel_cult_score}}{double. DESCRIPTION.}
  \item{\code{prest_cult_score}}{double. DESCRIPTION.}
  \item{\code{dom_cult_score}}{double. DESCRIPTION.}
}}
\usage{
leader_cult
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

