% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadershipdata.R
\docType{data}
\name{leader_text}
\alias{leader_text}
\title{Text-level data from HRAF variables}
\format{A data frame with 1212 rows and 51 variables:
\describe{
  \item{\code{cs_ID}}{integer. Code sheet ID.}
  \item{\code{cs_textrec_ID}}{double. Text record ID. In some cases one text record produced multiple code sheets.}
  \item{\code{text_type}}{integer. Classifying text type as: Cultural model, Case, or Both}
  \item{\code{c_culture_code}}{integer. OWC Culture code from eHRAF.}
  \item{\code{c_name}}{integer. HRAF culture name.}
  \item{\code{c_region}}{integer. HRAF culture region.}
  \item{\code{c_subregion}}{integer. HRAF sub-region.}
  \item{\code{c_subsistence_type}}{integer. HRAF subsistence type classification.}
  \item{\code{c_subsistence_code}}{integer. Numeric code assigned to HRAF subsistence classification.}
  \item{\code{demo_sex}}{integer. Sex of the leader described coded as male, female, or both.}
  \item{\code{dom_aggression}}{double. Does the leader use coercive control, aggression, or threats of violence? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{dom_assert.authority}}{double. Does the leader asserts authority over unwilling others, or have power over followers? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against.}
  \item{\code{dom_avoid.dom}}{double. Does the leader avoid being controlled by others? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{dom_fear}}{double. Does the leader instill fear in followers? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{dom_fighting}}{double. Does the leader demonstrate, or is known for superior fighting ability? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{dom_personality}}{double. Does the leader have a forceful or dominant personality? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{dom_reputation}}{double. Does the leader's reputation lead followers to be submissive? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{dom_strong}}{double. Is the leader seen as physically stronger than most followers? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{hooper_performance}}{double. asf? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{hooper_sanction.freeriders}}{double. asf? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{hooper_payoff}}{double. asf? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{hooper_group.size}}{double. asf? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{hooper_coop.activities}}{double. asf? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{hooper_coop.activities}}{double. asf? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{neel_better.mates}}{double. Does the leader have higher quality mates relative to followers? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{neel_big.family}}{double. Does the leader have a larger (nuclear) family size relative to followers? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{neel_intelligence}}{double. Does the leader have high intelligence, above average knowledge, high cognitive abilities? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{neel_polygynous}}{double. Is the leader polygynous? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_counsel}}{double. Does the leader provide valued opinions, or is sought out for counsel? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_emulated}}{double. Do followers emulate the leader? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_expertise}}{double. Does the leader demonstrate expert knowledge, or have unique talents and abilities? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_family}}{double. Does the leader gain prestige or influence based on family level prestige? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_f.exp.success}}{double. Do followers have an expectation of success of the leader? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_likable}}{double. Is the leader likable? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{prestige_respected}}{double. Is the leader respected, admired, or held in high esteem? Coded ‘1’ for evidence for, ‘0’ for no evidence, and ‘-1’ for evidence against. }
  \item{\code{doc_ID}}{double. asf.}
  \item{\code{doc_field_date}}{double. asf.}
  \item{\code{doc_coverage_date}}{double. asf.}
  \item{\code{doc_evaluation}}{double. asf.}
  \item{\code{doc_field_date}}{double. asf.}
  \item{\code{author_ID}}{double. asf.}
  \item{\code{evidence_prestige_for}}{double. The sum of evidence for the prestige model, e.g., the number of prestige-model variables coded '1' for the codesheet.}
  \item{\code{evidence_prestige_against}}{double. The sum of evidence against the prestige model, e.g., the number of prestige-model variables coded '-1' for the codesheet.}
  \item{\code{evidence_dom_for}}{double. The sum of evidence for the dominance model, e.g., the number of dominance-model variables coded '1' for the codesheet.}
  \item{\code{evidence_dom_against}}{double. The sum of evidence against the dominance model, e.g., the number of dominance-model variables coded '-1' for the codesheet.}
  \item{\code{evidence_neel_for}}{double. The sum of evidence for the neel model, e.g., the number of neel-model variables coded '1' for the codesheet.}
  \item{\code{evidence_neel_against}}{double. The sum of evidence against the neel model, e.g., the number of neel-model variables coded '-1' for the codesheet.}
  \item{\code{evidence_hooper_for}}{double. The sum of evidence for the Hooper model, e.g., the number of hooper-model variables coded '1' for the codesheet.}
  \item{\code{evidence_hooper_against}}{double. The sum of evidence against the Hooper model, e.g., the number of hooper-model variables coded '-1' for the codesheet.}
  \item{\code{subsistence}}{integer. Recoded HRAF subsistence classification with the value lables of: “hunter-gatherers”, “pastoralists”, “horticulturalists”, “agriculturalists”, and “other”}
  \item{\code{region}}{integer. SCCS region classification using the following value lables: "Africa", "Circum-Mediterranean", "East Eurasia", "Insular Pacific", "North America", "South America"}
  \item{\code{dom_against}}{double. The number of dominance model variables evidence against from the codesheet.}
  \item{\code{dom_none}}{double. The number of dominance model variables providing no evidence from the codesheet.}
  \item{\code{dom_sum}}{double. The number of dominance model variables that provide any evidence from that codesheet.}
  \item{\code{prestige_none}}{double. The number of prestige model variables providing no evidence from the codesheet.}
  \item{\code{prestige_sum}}{double. The number of prestige model variables that provide any evidence from that codesheet.}
  \item{\code{prestige_against}}{double. The number of prestige model variables evidence against from the codesheet.}
  \item{\code{neel_none}}{double. The number of neel model variables providing no evidence from the codesheet.}
  \item{\code{neel_sum}}{double. The number of neel model variables that provide any evidence from that codesheet.}
  \item{\code{neel_against}}{double. The number of Neel model variables evidence against from the codesheet.}
  \item{\code{hooper_none}}{double. The number of Hooper model variables providing no evidence from the codesheet.}
  \item{\code{hooper_sum}}{double. The number of Hooper model variables that provide any evidence from that codesheet.}
  \item{\code{hooper_against}}{double. The number of Hooper model variables evidence against from the codesheet.}
  \item{\code{neel_totals}}{double. The sum of all Neel model variables from that codesheet, e.g., '1's + '-1's.}
  \item{\code{hooper_totals}}{double. The sum of all Hooper model variables from that codesheet, e.g., '1's + '-1's.}
  \item{\code{dom_totals}}{double. The sum of all dominance model variables from that codesheet, e.g., '1's + '-1's.}
  \item{\code{prest_totals}}{double. The sum of all prestige model variables from that codesheet, e.g., '1's + '-1's.}
}}
\source{
\url{http://www.ehrafsite.info/}
}
\usage{
leader_text
}
\description{
1200+ ethnographic text coded as evidence for, against, no evidence on variables from theories of leadership.
}
\keyword{datasets}

